+++ src/components/ChannelPreview.js
@@ -15,10 +15,26 @@ export class ChannelPreview extends PureComponent {
   }
 
   static propTypes = {
+    /** **Available from [chat context](https://getstream.github.io/stream-chat-react/#chat)** */
     channel: PropTypes.object.isRequired,
+    /** Current selected channel object */
     activeChannel: PropTypes.object.isRequired,
+    /** Setter for selected channel */
     setActiveChannel: PropTypes.func.isRequired,
+    /**
+     * Available built-in options (also accepts the same props as):
+     *
+     * 1. [ChannelPreviewCompact](https://getstream.github.io/stream-chat-react/#ChannelPreviewCompact) (default)
+     * 2. [ChannelPreviewLastMessage](https://getstream.github.io/stream-chat-react/#ChannelPreviewLastMessage)
+     * 3. [ChannelPreviewMessanger](https://getstream.github.io/stream-chat-react/#ChannelPreviewMessanger)
+     *
+     * The Preview to use, defaults to ChannelPreviewLastMessage
+     * */
     Preview: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
+    /**
+     * Object containing watcher parameters
+     * @see See [Pagination documentation](https://getstream.io/chat/docs/#channel_pagination) for a list of available fields for sort.
+     * */
     watchers: PropTypes.object,
   };
 
@@ -33,11 +49,15 @@ export class ChannelPreview extends PureComponent {
 
     this.setState({ unread });
     channel.on('message.new', this.handleEvent);
+    channel.on('message.updated', this.handleEvent);
+    channel.on('message.deleted', this.handleEvent);
   }
 
   componentWillUnmount() {
     const channel = this.props.channel;
     channel.off('message.new', this.handleEvent);
+    channel.off('message.updated', this.handleEvent);
+    channel.off('message.deleted', this.handleEvent);
   }
 
   handleEvent = (event) => {