+++ src/components/TypingIndicator.js
@@ -1,8 +1,16 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
 import { Avatar } from './Avatar';
 
 export class TypingIndicator extends React.PureComponent {
+  static propTypes = {
+    /** @see See [chat context](https://getstream.github.io/stream-chat-react/#chatcontext) doc */
+    client: PropTypes.object,
+    /** @see See [channel context](https://getstream.github.io/stream-chat-react/#channelcontext) doc */
+    typing: PropTypes.object,
+  };
+
   render() {
     // if (Object.keys(this.props.typing).length <= 0) {
     //   return null;